refactor(execution): replace deep imports with public API imports (#2445)
Replace non-public API deep imports in execution module with public API
imports from the canonical entry points (models, clients, exceptions).
